# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Travel Agency Documentation",
        "description": "",
        "contact": {
            "email": "rahmanimahdi16@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Demo API Server"
        }
    ],
    "paths": {
        "/api/login": {
            "post": {
                "tags": [
                    "Users"
                ],
                "description": "login the user",
                "operationId": "LoginUser",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LoginData"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "unauthorized"
                    }
                }
            }
        },
        "/api/travels/{travel}/tours": {
            "get": {
                "tags": [
                    "Tours"
                ],
                "summary": "Get list of tours",
                "description": "Returns list of tours",
                "operationId": "getTravelToursList",
                "parameters": [
                    {
                        "name": "travel",
                        "in": "path",
                        "description": "slug of the Travel",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "test"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/{travel}/tours": {
            "post": {
                "tags": [
                    "Tours"
                ],
                "description": "store new Tour",
                "operationId": "storePartDelivery",
                "parameters": [
                    {
                        "name": "travel",
                        "in": "path",
                        "description": "travel_id of the Travel",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": "1"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TourData"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "unauthorized"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/travels": {
            "get": {
                "tags": [
                    "Travels"
                ],
                "summary": "Get list of travels",
                "description": "Returns list of travels",
                "operationId": "getTravelsList",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "TravelParts"
                ],
                "description": "store new travel request",
                "operationId": "storeTravel",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TravelData"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "403": {
                        "description": "unauthorized"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/travels/{travel}/update": {
            "put": {
                "tags": [
                    "TravelParts"
                ],
                "description": "update delivery part",
                "operationId": "updateTravel",
                "parameters": [
                    {
                        "name": "travel",
                        "in": "path",
                        "description": "travel_id of the Travel",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TravelData"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "unauthorized"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "LoginData": {
                "title": "Login data",
                "description": "login Data",
                "required": [
                    "email",
                    "password"
                ],
                "properties": {
                    "email": {
                        "title": "email",
                        "description": "email of the user",
                        "type": "string",
                        "example": "test@example.com"
                    },
                    "password": {
                        "title": "password",
                        "description": "password of the User",
                        "type": "string",
                        "example": "test"
                    }
                },
                "type": "object"
            },
            "TourData": {
                "title": "Tour data",
                "description": "tour Data",
                "required": [
                    "name",
                    "starting_date",
                    "ending_date",
                    "price"
                ],
                "properties": {
                    "name": {
                        "title": "name",
                        "description": "name of the new Travel",
                        "type": "string",
                        "example": "test"
                    },
                    "starting_date": {
                        "title": "starting_date",
                        "description": "starting_date of the new Tour",
                        "type": "string",
                        "example": "2023_01_01"
                    },
                    "ending_date": {
                        "title": "ending_date",
                        "description": "nending_date of the new Tour",
                        "type": "integer",
                        "example": "2023_01_01"
                    },
                    "price": {
                        "title": "price",
                        "description": "price of the new Tour",
                        "type": "integer",
                        "example": "100"
                    }
                },
                "type": "object"
            },
            "TravelData": {
                "title": "Travel data",
                "description": "travel Data",
                "required": [
                    "is_public",
                    "name",
                    "description",
                    "number_of_days"
                ],
                "properties": {
                    "is_public": {
                        "title": "is_public",
                        "description": "is_public of the new Travel",
                        "type": "boolean",
                        "example": "true"
                    },
                    "name": {
                        "title": "name",
                        "description": "name of the new Travel",
                        "type": "string",
                        "example": "test"
                    },
                    "description": {
                        "title": "description",
                        "description": "description of the new Travel",
                        "type": "string",
                        "example": "test"
                    },
                    "number_of_days": {
                        "title": "number_of_days",
                        "description": "number_of_days of the new Travel",
                        "type": "integer",
                        "example": "1"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "tags": [
        {
            "name": "Travel Agency",
            "description": "API Endpoints of Travel Agency"
        }
    ]
}